+++ src/Controller/UserController.php
@@ -157,13 +157,17 @@ class UserController extends AbstractController
      */
     public function delete(Request $request, User $user): Response
     {
-        if ($this->isCsrfTokenValid('delete' . $user->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($user);
-            $entityManager->flush();
-            $this->addFlash('danger', 'Vartotojas sėkmingai pašalintas');
+        if ($user->getId() === $this->get('security.token_storage')->getToken()->getUser()->getId()) {
+            if ($this->isCsrfTokenValid('delete' . $user->getId(), $request->request->get('_token'))) {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->remove($user);
+                $entityManager->flush();
+                $this->addFlash('danger', 'Vartotojas sėkmingai pašalintas');
+            }
+
+            return $this->redirectToRoute('user_index');
         }
 
-        return $this->redirectToRoute('user_index');
+        throw $this->createNotFoundException('You are not allowed to reach this site.');
     }
 }
